+++ Assets/Scenes/World.unity
@@ -14,7 +14,7 @@ OcclusionCullingSettings:
 RenderSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 9
-  m_Fog: 1
+  m_Fog: 0
   m_FogColor: {r: 0.5294118, g: 0.80784315, b: 0.92156863, a: 1}
   m_FogMode: 1
   m_FogDensity: 0.01
