--- Packages/io.codeblaze.vloxyengine/Runtime/Engine/Components/ChunkBehaviourPool.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: b32ffd2e9166416d908e7aa5be404898
-timeCreated: 1605593190
