+++ Packages/io.codeblaze.vloxyengine/Runtime/Engine/Data/ChunkManager.cs
@@ -13,7 +13,6 @@ namespace CodeBlaze.Vloxy.Engine.Data {
     public class ChunkManager {
 
         internal ChunkStore Store { get; }
-        internal ChunkState State { get; }
         
         private ChunkSettings _ChunkSettings;
         
@@ -23,7 +22,6 @@ namespace CodeBlaze.Vloxy.Engine.Data {
         public ChunkManager(VloxySettings settings) {
             _ChunkSettings = settings.Chunk;
 
-            State = new ChunkState(settings);
             Store = new ChunkStore(settings);
             
             var viewRegionSize = _ChunkSettings.DrawDistance.CubedSize();
@@ -86,7 +84,6 @@ namespace CodeBlaze.Vloxy.Engine.Data {
                     for (int y = -size; y <= size; y++) {
                         var position = new int3(x, y, z) * _ChunkSettings.ChunkSize;
                         result[index] = position;
-                        State.SetState(position, ChunkState.State.STREAMING);
                         index++;
                     }
                 }
