--- Packages/io.codeblaze.vloxyengine/Runtime/Engine/Data/ChunkState.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 62135b1997bb4420a4b31a62cc585731
-timeCreated: 1654057402
