--- Packages/io.codeblaze.vloxyengine/Runtime/Engine/Jobs/Data/ChunkDataSchedulerV2.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a163015120e04be2843882977f86ae3a
-timeCreated: 1679376174
