+++ Packages/io.codeblaze.vloxyengine/Runtime/Engine/Jobs/Mesh/MeshBuildScheduler.cs.meta
@@ -1,3 +1,3 @@
-﻿fileFormatVersion: 2
-guid: 39a8a766125f4a09a18378c5236e251c
-timeCreated: 1638904496
+fileFormatVersion: 2
+guid: ba97fd14d11347a09eb487b45a69bbc7
+timeCreated: 1679366933
