--- Packages/io.codeblaze.vloxyengine/Runtime/Engine/Jobs/Mesh/MeshBuildSchedulerV2.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: ba97fd14d11347a09eb487b45a69bbc7
-timeCreated: 1679366933
