+++ Packages/io.codeblaze.vloxyengine/Runtime/Engine/Settings/SchedulerSettings.cs
@@ -13,6 +13,9 @@ namespace CodeBlaze.Vloxy.Engine.Settings {
         [HideInInspector]
         public int StreamingBatchSize;
 
+        [Tooltip("Framerate at which the scheduler updates")]
+        public int TickRate = 16;
+
     }
 
 }
