+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,6 @@ EditorBuildSettings:
     path: Assets/Scenes/Main.unity
     guid: c4f88d7efa01d2a48ac2b95518ae254e
   - enabled: 1
-    path: Packages/io.codeblaze.vloxyengine/Samples/Textured/TexturedWorld.unity
-    guid: ec02e76261abb6d46a6898c3bddf3bb1
+    path: Assets/Scenes/World.unity
+    guid: 2e302767699d543849139906093e207a
   m_configObjects: {}
